Release/v0.1.5

+++ src/dependency_downloader.rs
@@ -8,7 +8,6 @@ use zip_extract::ZipExtractError;
 
 use crate::config::{ Dependency, read_config };
 use crate::utils::get_current_working_dir;
-use crate::FOUNDRY;
 
 // TODOs:
 // - needs to be downloaded in parallel
@@ -16,11 +15,16 @@ pub async fn download_dependencies(
     dependencies: &Vec<Dependency>,
     clean: bool
 ) -> Result<(), DownloadError> {
+    // clean dependencies folder if flag is true
     if clean {
         let dep_path = get_current_working_dir().unwrap().join("dependencies");
-        fs::remove_dir_all(&dep_path).unwrap();
-        fs::create_dir(&dep_path).unwrap();
+        if dep_path.is_dir() {
+            fs::remove_dir_all(&dep_path).unwrap();
+            fs::create_dir(&dep_path).unwrap();
+        }
     }
+
+    // downloading dependencies to dependencies folder
     for dependency in dependencies.iter() {
         let file_name: String = format!("{}-{}.zip", dependency.name, dependency.version);
         match download_dependency(&file_name, &dependency.url).await {
@@ -52,8 +56,7 @@ pub fn unzip_dependencies(dependencies: &Vec<Dependency>) -> Result<(), ZipExtra
 pub async fn download_dependency_remote(
     dependency_name: &String,
     dependency_version: &String,
-    remote_url: &String,
-    foundry_setup: &FOUNDRY
+    remote_url: &String
 ) -> Result<String, DownloadError> {
     let res: Response = get(format!("{}", remote_url)).await.unwrap();
     let body: String = res.text().await.unwrap();
@@ -61,7 +64,9 @@ pub async fn download_dependency_remote(
         .unwrap()
         .join(".dependency_reading.toml");
     fs::write(&tmp_path, body).expect("Unable to write file");
-    let dependencies: Vec<Dependency> = read_config((&tmp_path).to_str().unwrap().to_string(), foundry_setup);
+    let dependencies: Vec<Dependency> = read_config(
+        (&tmp_path).to_str().unwrap().to_string()
+    );
     for dependency in dependencies.iter() {
         if dependency.name == *dependency_name && dependency.version == *dependency_version {
             match
